+++ hippodamus/src/dd/kms/hippodamus/impl/coordinator/configuration/ExecutionCoordinatorBuilderImpl.java
@@ -3,7 +3,6 @@ package dd.kms.hippodamus.impl.coordinator.configuration;
 import dd.kms.hippodamus.api.coordinator.ExecutionCoordinator;
 import dd.kms.hippodamus.api.coordinator.TaskType;
 import dd.kms.hippodamus.api.coordinator.configuration.ExecutionCoordinatorBuilder;
-import dd.kms.hippodamus.api.coordinator.configuration.WaitMode;
 import dd.kms.hippodamus.api.logging.LogLevel;
 import dd.kms.hippodamus.api.logging.Logger;
 import dd.kms.hippodamus.impl.coordinator.ExecutionCoordinatorImpl;
@@ -19,7 +18,7 @@ public class ExecutionCoordinatorBuilderImpl extends CoordinatorBuilderBase<Exec
 	}
 
 	@Override
-	ExecutionCoordinator createCoordinator(Map<TaskType, ExecutorServiceWrapper> executorServiceWrappersByTaskType, Logger logger, LogLevel minimumLogLevel, boolean verifyDependencies, WaitMode waitMode) {
-		return new ExecutionCoordinatorImpl(executorServiceWrappersByTaskType, logger, minimumLogLevel, verifyDependencies, waitMode);
+	ExecutionCoordinator createCoordinator(Map<TaskType, ExecutorServiceWrapper> executorServiceWrappersByTaskType, Logger logger, LogLevel minimumLogLevel, boolean verifyDependencies) {
+		return new ExecutionCoordinatorImpl(executorServiceWrappersByTaskType, logger, minimumLogLevel, verifyDependencies);
 	}
 }